Added Jing AI, adjusted picture format, new title, changes page styling
As stated in the above, I've added a lot of styling changes to help format the page better, including converting the top part into an image so it can scale well on mobile etc. I have added a new person and cropped the images so that their heads are towards the centre, which looks better on most formats. Finally, I've added some SEO content to help on that side of things etc...