Final Adjustments & added more Speakers
Final formatting adjustments, such as the size of the speaker boxes, the colours they turn into on hover, more speakers being added in the speaker json and changing the formatting for each png (and ensure all the ones have been added and named correctly). Also minor background change to make it look a bit better with a slightly different gradient.